Update WordCounter.java
In this code, we first prompt the user to enter text or provide a file path. If the input is a valid file path, we read the content of the file; otherwise, we use the provided text. We then split the text into words using space or punctuation as delimiters and count the number of words while ignoring common words.

The code also provides additional statistics like the number of unique words and their frequencies. Common words, which are often considered insignificant for analysis, are ignored in the word count.